tools.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()

def get_weather(city: str) -> str:
    """Get current weather for a city using OpenWeatherMap API."""
    api_key = os.getenv("OPENWEATHER_API_KEY")
    if not api_key:
        return "Error: OpenWeatherMap API key not found in environment variables"
    
    base_url = "http://api.openweathermap.org/data/2.5/weather"
    params = {
        "q": city,
        "appid": api_key,
        "units": "metric"  # For Celsius
    }
    
    try:
        response = requests.get(base_url, params=params)
        response.raise_for_status()
        data = response.json()
        
        weather_desc = data["weather"][0]["description"]
        temp = data["main"]["temp"]
        humidity = data["main"]["humidity"]

        print(f"Current weather in {city}: {weather_desc}. Temperature: {temp}°C, Humidity: {humidity}%")
        
        return f"Current weather in {city}: {weather_desc}. Temperature: {temp}°C, Humidity: {humidity}%"
    except Exception as e:
        return f"Error getting weather data: {str(e)}"