requirements.txt
openai
python-dotenv==1.0.0
requests==2.31.0
Flask==2.1.2
